# backend/docs/openapi.yml
openapi: 3.1.0
info:
  title: EdTech OS — Multi-Tenant Coaching API
  version: 1.0.0
  description: |
    **Multi-Tenant EdTech OS** — complete REST API for coaching institute management.

    ### Authentication
    All protected routes require a JWT Bearer token obtained from `POST /auth/login`.
    Include it as: `Authorization: Bearer <accessToken>`

    ### Roles
    | Role | Description |
    |------|-------------|
    | `super_admin` | Platform owner — manages all institutes |
    | `coaching_admin` | Institute owner — manages students, teachers, fees |
    | `teacher` | Takes attendance, uploads content, creates live classes |
    | `student` | Views content, fees, live classes |

    ### Tenant Isolation
    Every request is scoped to the authenticated user's `tenantId` from their JWT.
    Never pass `tenantId` in the body — it is always read from the token.

  contact:
    name: EdTech OS Support
    email: support@edtechos.com
  license:
    name: MIT
servers:
  - url: http://localhost:4000/api/v1
    description: Local development server
  - url: https://api.edtechos.com/api/v1
    description: Production server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token from POST /auth/login

  # ───────────────────────────────────────────────────────────────────────────
  # REUSABLE SCHEMAS
  # ───────────────────────────────────────────────────────────────────────────
  schemas:
...
    Error:
      type: object
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Phone must be 10-15 digits
        errors:
          type: array
          items:
            type: object
...
    CreateTeacherRequest:
      type: object
      required: [fullName, phone, password]
      properties:
        fullName: { type: string, example: "Sunita Patil" }
        phone:    { type: string, example: "919000000012" }
        password: { type: string, example: "Teacher@123", minLength: 6 }
        email:    { type: string, format: email, example: "sunita@apexacademy.com" }
    TeacherItem:
      type: object
      properties:
        id:       { type: integer, example: 4 }
        fullName: { type: string, example: "Sunita Patil" }
        phone:    { type: string, example: "919000000012" }
        email:    { type: string, nullable: true }

    # ── Student ──────────────────────────────────────────────────────────────
    CreateStudentRequest:
      type: object
      required: [fullName, phone, password, parentPhone, batchId]
      properties:
        fullName:    { type: string, example: "Aarav Joshi" }
        phone:       { type: string, example: "919000000013" }
        password:    { type: string, example: "Student@123", minLength: 6 }
        parentName:  { type: string, example: "Ramesh Joshi" }
        parentPhone: { type: string, example: "919000000099" }
        grade:       { type: string, example: "Class 11" }
        rollNo:      { type: string, example: "APX-001" }
        batchId:     { type: integer, example: 3 }
    StudentListItem:
      type: object
      properties:
        id:          { type: integer, example: 7 }
        fullName:    { type: string, example: "Aarav Joshi" }
        rollNo:      { type: string, nullable: true, example: "APX-001" }
        grade:       { type: string, nullable: true, example: "Class 11" }
        parentPhone: { type: string, example: "919000000099" }

    # ── Batch & Subject ───────────────────────────────────────────────────────
    CreateBatchRequest:
      type: object
      required: [name]
      properties:
        name:  { type: string, example: "JEE 2026 Morning" }
        grade: { type: string, example: "Class 11" }
    BatchItem:
      type: object
      properties:
        id:           { type: integer, example: 3 }
        name:         { type: string, example: "JEE 2026 Morning" }
        grade:        { type: string, nullable: true, example: "Class 11" }
        studentCount: { type: integer, example: 48 }
    CreateSubjectRequest:
      type: object
      required: [name]
      properties:
        name: { type: string, example: Physics }
    SubjectItem:
      type: object
      properties:
        id:   { type: integer, example: 1 }
        name: { type: string, example: Physics }

    # ── Timetable ────────────────────────────────────────────────────────────
    CreateTimetableRequest:
      type: object
      required: [batchId, teacherId, dayOfWeek, startTime, endTime]
      properties:
        batchId:    { type: integer, example: 3 }
        subjectId:  { type: integer, example: 1 }
        teacherId:  { type: integer, example: 4 }
        dayOfWeek:  { type: integer, minimum: 0, maximum: 6, example: 1, description: "0=Sun, 1=Mon … 6=Sat" }
        startTime:  { type: string, example: "09:00" }
        endTime:    { type: string, example: "10:30" }
    TimetableItem:
      type: object
      properties:
        id:        { type: integer, example: 12 }
        batch:     { type: string, example: "JEE 2026 Morning" }
        subject:   { type: string, nullable: true, example: Physics }
        teacher:   { type: string, example: "Sunita Patil" }
        startTime: { type: string, example: "09:00:00" }
        endTime:   { type: string, example: "10:30:00" }

    # ── Fees ─────────────────────────────────────────────────────────────────
    CreateFeeStructureRequest:
      type: object
      required: [title, amount]
      properties:
        batchId: { type: integer, example: 3 }
        title:   { type: string, example: "Term 1 Fee" }
        amount:  { type: integer, example: 15000 }
        dueDate: { type: string, format: date, example: "2026-08-15" }
    RecordPaymentRequest:
      type: object
      required: [studentId, amountPaid]
      properties:
        studentId:       { type: integer, example: 7 }
        feeStructureId:  { type: integer, example: 2 }
        amountPaid:      { type: integer, example: 15000 }
        method:          { type: string, enum: [cash, upi, card], example: upi }
    FeeListItem:
      type: object
      properties:
        studentId: { type: integer, example: 7 }
        name:      { type: string, example: "Aarav Joshi" }
        total:     { type: integer, example: 15000 }
        paid:      { type: integer, example: 10000 }
        pending:   { type: integer, example: 5000 }

    # ── Attendance ───────────────────────────────────────────────────────────
...
    BroadcastResponse:
      type: object
      properties:
        recipientCount:
          type: integer
          example: 12
          description: Number of users a notification row was inserted for.
...
    BroadcastToStudentsRequest:
      type: object
      required: [title]
      properties:
        title:   { type: string, example: "Holiday tomorrow" }
        body:    { type: string, example: "Institute will remain closed tomorrow for a public holiday." }
        batchId:
          type: integer
          example: 4
          description: Omit to broadcast to every student in the institute; set to target only one batch.

  # ── Common Responses ───────────────────────────────────────────────────────
  responses:
...
    Forbidden:
      description: Insufficient role or plan
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          examples:
            role_forbidden:
              value: { code: FORBIDDEN, message: "Access denied for your role" }
            trial_expired:
              value: { code: TRIAL_EXPIRED, message: "Your free trial has ended. Please subscribe." }
            plan_upgrade:
              value: { code: PLAN_UPGRADE_REQUIRED, message: "This feature requires the Pro plan or higher." }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { code: NOT_FOUND, message: "Resource not found" }
    ValidationError:
      description: Request body / params failed validation
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { code: VALIDATION_ERROR, message: "phone: Phone must be 10-15 digits" }

# ─────────────────────────────────────────────────────────────────────────────
# PATHS
# ─────────────────────────────────────────────────────────────────────────────
paths:
...
  /admin/teachers:
    get:
      tags: [Coaching Admin — Teachers]
      summary: List teachers
      operationId: listTeachers
      security:
        - BearerAuth: []
      responses:
        '200':
          description: All teachers
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/TeacherItem' }
    post:
      tags: [Coaching Admin — Teachers]
      summary: Add a teacher
      description: |
        **Requires Elite plan** (`teacher_accounts` feature).
        Also blocked when subscription is expired/past_due.
      operationId: createTeacher
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateTeacherRequest' }
      responses:
        '201':
          description: Teacher created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TeacherItem' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /admin/teachers/{id}:
    delete:
      tags: [Coaching Admin — Teachers]
      summary: Remove a teacher
      operationId: deleteTeacher
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, example: 4 }
      responses:
        '200':
          description: Teacher removed
          content:
            application/json:
              example: { success: true }
        '404': { $ref: '#/components/responses/NotFound' }

  # ══════════════════════════════════════════════════════════════════════════
  # COACHING ADMIN — STUDENTS
  # ══════════════════════════════════════════════════════════════════════════
  /admin/students:
    get:
      tags: [Coaching Admin — Students]
      summary: List students
      operationId: listStudents
      security:
        - BearerAuth: []
      parameters:
        - name: batchId
          in: query
          schema: { type: integer, example: 3 }
          description: Filter by batch
      responses:
        '200':
          description: Student list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/StudentListItem' }
    post:
      tags: [Coaching Admin — Students]
      summary: Enroll a new student
      description: Creates user account + student profile + batch enrollment in one transaction.
      operationId: createStudent
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateStudentRequest' }
      responses:
        '201':
          description: Student enrolled
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StudentListItem' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /admin/students/{id}:
    delete:
      tags: [Coaching Admin — Students]
      summary: Remove a student
      operationId: deleteStudent
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, example: 7 }
      responses:
        '200':
          description: Student removed
          content:
            application/json:
              example: { success: true }

  # ══════════════════════════════════════════════════════════════════════════
  # COACHING ADMIN — BATCHES & SUBJECTS
  # ══════════════════════════════════════════════════════════════════════════
  /admin/batches:
    get:
      tags: [Coaching Admin — Batches & Subjects]
      summary: List batches
      operationId: listBatches
      security:
        - BearerAuth: []
      responses:
        '200':
          description: All batches
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/BatchItem' }
    post:
      tags: [Coaching Admin — Batches & Subjects]
      summary: Create a batch
      operationId: createBatch
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateBatchRequest' }
      responses:
        '201':
          description: Batch created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BatchItem' }
  /admin/subjects:
    get:
      tags: [Coaching Admin — Batches & Subjects]
      summary: List subjects
      operationId: listSubjects
      security:
        - BearerAuth: []
      responses:
        '200':
          description: All subjects
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/SubjectItem' }
    post:
      tags: [Coaching Admin — Batches & Subjects]
      summary: Create a subject
      operationId: createSubject
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateSubjectRequest' }
      responses:
        '201':
          description: Subject created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SubjectItem' }

  # ══════════════════════════════════════════════════════════════════════════
  # COACHING ADMIN — TIMETABLE
  # ══════════════════════════════════════════════════════════════════════════
  /admin/timetable:
    get:
      tags: [Coaching Admin — Timetable]
      summary: List timetable entries
      operationId: listTimetable
      security:
        - BearerAuth: []
      parameters:
        - name: day
          in: query
          schema: { type: integer, minimum: 0, maximum: 6, example: 1 }
          description: Filter by day of week (0=Sun…6=Sat)
      responses:
        '200':
          description: Timetable entries
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/TimetableItem' }
    post:
      tags: [Coaching Admin — Timetable]
      summary: Add a timetable entry (teacher allocation)
      operationId: createTimetable
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateTimetableRequest' }
      responses:
        '201':
          description: Timetable entry created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TimetableItem' }

  # ══════════════════════════════════════════════════════════════════════════
  # COACHING ADMIN — FEES
  # ══════════════════════════════════════════════════════════════════════════
  /admin/fees:
    get:
      tags: [Coaching Admin — Fees]
      summary: Fee collection overview
      operationId: listFees
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          schema: { type: string, enum: [pending, paid, all], example: pending }
      responses:
        '200':
          description: Fee list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/FeeListItem' }
  /admin/fees/structures:
    post:
      tags: [Coaching Admin — Fees]
      summary: Create a fee structure
      operationId: createFeeStructure
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateFeeStructureRequest' }
      responses:
        '201':
          description: Fee structure created
          content:
            application/json:
              example: { id: 2, title: "Term 1 Fee", amount: 15000, dueDate: "2026-08-15" }
  /admin/fees/payments:
    post:
      tags: [Coaching Admin — Fees]
      summary: Record a fee payment
      operationId: recordPayment
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RecordPaymentRequest' }
      responses:
        '201':
          description: Payment recorded
          content:
            application/json:
              example:
                payment: { id: 18, amountPaid: 15000, paidOn: "2026-07-09", method: upi }
                receiptNo: "APX-2026-00018"
  /admin/fees/{studentId}/remind:
    post:
      tags: [Coaching Admin — Fees]
      summary: Get WhatsApp fee reminder link for a student
      description: |
        **Requires Pro or Elite plan** (`whatsapp_reminders` feature).
        Returns a pre-filled `wa.me` URL to send a fee reminder to the student's parent.
      operationId: feeReminder
      security:
        - BearerAuth: []
      parameters:
        - name: studentId
          in: path
          required: true
          schema: { type: integer, example: 7 }
      responses:
        '200':
          description: WhatsApp reminder URL
          content:
            application/json:
              example:
                waUrl: "https://wa.me/919000000099?text=Namaste%20ji%2C%20Aarav%20ki%20Term%201%20fees%20%E2%82%B95000%20pending%20hai."
        '403': { $ref: '#/components/responses/Forbidden' }

  # ══════════════════════════════════════════════════════════════════════════
  # COACHING ADMIN — BRANDING
  # ══════════════════════════════════════════════════════════════════════════
  /admin/branding:
    get:
      tags: [Coaching Admin — Branding]
      summary: Get institute branding
      operationId: getBranding
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Branding info
          content:
            application/json:
              example: { logoUrl: "https://cdn.edtechos.com/apex.png", primaryColor: "#2563EB", name: "Apex Academy" }
    put:
      tags: [Coaching Admin — Branding]
      summary: Update branding (logo + color)
      description: |
        **Requires Elite plan** (`custom_branding` feature).
        Also blocked when subscription is expired.
      operationId: updateBranding
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                logoUrl:      { type: string, format: uri, example: "https://cdn.edtechos.com/apex-new.png" }
                primaryColor: { type: string, example: "#16A34A" }
      responses:
        '200':
          description: Branding updated
          content:
            application/json:
              example: { logoUrl: "https://cdn.edtechos.com/apex-new.png", primaryColor: "#16A34A" }
        '403': { $ref: '#/components/responses/Forbidden' }

  # ══════════════════════════════════════════════════════════════════════════
  # COACHING ADMIN — REPORTS
  # ══════════════════════════════════════════════════════════════════════════
  /admin/reports/performance:
    get:
      tags: [Coaching Admin — Reports]
      summary: Institute performance report
      description: |
        **Requires Pro or Elite plan** (`performance_reports` feature).
        Returns average attendance, average marks, and top students.
      operationId: performanceReport
      security:
        - BearerAuth: []
      parameters:
        - name: batchId
          in: query
          schema: { type: integer, example: 3 }
          description: Filter by batch (optional)
      responses:
        '200':
          description: Performance report
          content:
            application/json:
              example:
                avgAttendance: 87.4
                avgMarks: 72.1
                topStudents:
                  - studentId: 7
                    name: "Aarav Joshi"
                    avgMarks: 91.5
                    attendancePct: 96.0
        '403': { $ref: '#/components/responses/Forbidden' }

  # ══════════════════════════════════════════════════════════════════════════
  # COACHING ADMIN — NOTIFICATIONS
  # ══════════════════════════════════════════════════════════════════════════
...
  /admin/notifications/broadcast:
    post:
      tags: [Coaching Admin — Notifications]
      summary: Broadcast a notification to students
      description: |
        Inserts one notification row per matching student in the caller's own
        institute (tenant is always taken from the JWT, never the request
        body). Omit `batchId` to reach every student, or set it to target
        only one batch. Blocked when subscription is expired/past_due.
      operationId: broadcastToStudents
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BroadcastToStudentsRequest' }
            examples:
              all_students:
                summary: Broadcast to the whole institute
                value: { title: "Holiday tomorrow", body: "Institute closed for a public holiday." }
              by_batch:
                summary: Broadcast filtered by batch
                value: { title: "Test postponed", batchId: 4 }
      responses:
        '201':
          description: Broadcast sent
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BroadcastResponse' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/ValidationError' }

  # ══════════════════════════════════════════════════════════════════════════
  # TEACHER
  # ══════════════════════════════════════════════════════════════════════════
...